requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

eodag==2.11.0
folium==0.15.1
streamlit==1.33.0
streamlit-chat==0.1.1
streamlit-folium==0.18.0

jupyter-server-proxy
nbserverproxy
owslib
streamlit-shadcn-ui==0.1.18
streamlit-supabase-auth==1.0.2
supabase==2.4.1
stripe==8.8.0
python-dotenv
extra-streamlit-components
beautifulsoup4
streamlit-lightweight-charts